backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
llama-cpp-python>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.4.0
sse-starlette>=1.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
pypdf>=4.0.0
cryptography>=3.1
scikit-learn>=1.3.0
plotly>=5.18.0
